requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
croniter==0.3.30
dask==2.7.0
decorator==4.4.1
defusedxml==0.6.0
distributed==2.7.0
docker==4.1.0
entrypoints==0.3
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-Session==0.3.1
fsspec==0.5.2
graphviz==0.13.2
HeapDict==1.0.1
idna==2.8
importlib-metadata==0.23
ipdb==0.12.2
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
locket==0.2.0
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-oneofschema==2.0.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.2
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
pendulum==2.0.5
pexpect==4.7.0
pickleshare==0.7.5
prefect==0.7.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.5
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.9.0
pyrsistent==0.15.5
python-box==3.4.5
python-dateutil==2.8.1
python-dotenv==0.10.3
python-slugify==4.0.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.1.0
tabulate==0.8.5
tblib==1.5.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
toml==0.10.0
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zict==1.0.0
zipp==0.6.0
